rainbird: Move RainBirdValve to entity.py (C7411 pylint fix)
CoordinatorEntity-derived classes must live in the 'entity' module per
the home-assistant-enforce-class-module pylint rule (C7411).

- Create entity.py containing RainBirdValve
- valve.py now only contains async_setup_entry and imports RainBirdValve
